Make cache methods non-generic (#28)
